feat: MET-1888 add faced and types
MET-1765, MET-1766, MET-1767, MET-1771
