+++ src/index.ts
@@ -1,10 +1,12 @@
 import CherryGLInstance from './cherry/CherryGL';
 
+export * from './assets';
 export * from './types';
 export * from './constants';
+
 export * from './facade';
-export * from './assets';
+
 // IMPORTANT: Don't change this line it will automaticlly change the version after build!
-export const CherryGLVersion = '0.0.0';
+export const CherryGLVersion = '[VI]{version}[/VI]';
 
 export default CherryGLInstance;
