+++ src/types/facade/CherryObjectAnimations.ts
@@ -1,0 +1,1 @@
+export type CherryObjectAnimations = {};
