+++ src/types/facade/CherryObjectComponents.ts
@@ -1,0 +1,1 @@
+export type CherryObjectComponents = {};
