+++ src/types/facade/CherryObjectInfo.ts
@@ -1,0 +1,1 @@
+export type CherryObjectInfo = {};
