+++ src/types/facade/CherryObjectMeshes.ts
@@ -1,0 +1,12 @@
+import {
+  CherryMesh,
+  CherryMeshGroup,
+  CherryMeshMaterial,
+} from '../cherry/CherryMesh';
+
+export type CherryObjectMeshes = {
+  objectMeshes: CherryMesh[];
+  objectMeshMaterials: CherryMeshMaterial[];
+  objectMeshGroups: CherryMeshGroup[];
+  shaderTypes: any;
+};
