+++ src/types/facade/ScriptAssets.ts
@@ -1,0 +1,1 @@
+export type ScriptAssets = { [key: string]: string };
