+++ src/types/facade/ShaderParameterType.ts
@@ -1,0 +1,40 @@
+export type ShaderParameterType =
+  | 'use_pbr'
+  | 'ao_ratio'
+  | 'ao_texture'
+  | 'ao_texture_channel'
+  | 'metalness_ratio'
+  | 'metalness_texture'
+  | 'metalness_texture_channel'
+  | 'roughness_ratio'
+  | 'roughness_texture'
+  | 'roughness_texture_channel'
+  | 'albedo_ratio'
+  | 'albedo_texture'
+  | 'albedo_video'
+  | 'use_alpha_channel'
+  | 'emissive_ratio'
+  | 'emissive_texture'
+  | 'diffuse_ibl_ratio'
+  | 'specular_pbr_ratio'
+  | 'specular_ibl_ratio'
+  | 'normal_texture'
+  | 'normal_ratio'
+  | 'uv_animation'
+  | 'opacity_ratio'
+  | 'opacity_texture'
+  | 'opacity_texture_channel'
+  | 'ambient_ratio'
+  | 'ambient_texture'
+  | 'ambient_video'
+  | 'diffuse_ratio'
+  | 'diffuse_texture'
+  | 'specular_ratio'
+  | 'specular_texture'
+  | 'specular_power'
+  | 'normal_texture'
+  | 'normal_ratio'
+  | 'uv_animation'
+  | 'opacity_ratio'
+  | 'opacity_texture'
+  | 'opacity_texture_channel';
