+++ src/types/facade/ShaderType.ts
@@ -1,0 +1,1 @@
+export type ShaderType = 'PBR' | 'STANDARD';
