+++ src/types/facade/ShaderValueType.ts
@@ -1,0 +1,1 @@
+export type ShaderValueType = 'float' | 'file' | 'string' | 'vec3' | 'boolean';
